# update height field midphase