updates for the height field midphase

```
mjwarp-testspeed benchmark/apptronik_apollo/scene_hfield.xml --nworld=4096 --nconmax=3000000 --njmax=512 --keyframe=0
```

this pr
```
Summary for 4096 parallel rollouts

Total JIT time: 0.59 s
Total simulation time: 27.57 s
Total steps per second: 148,559
Total realtime factor: 742.80 x
Total time per step: 6731.33 ns
```

main
```
Summary for 4096 parallel rollouts

Total JIT time: 0.59 s
Total simulation time: 28.04 s
Total steps per second: 146,093
Total realtime factor: 730.47 x
Total time per step: 6844.95 ns
```

---

if `nconmax` is reduced:

```
mjwarp-testspeed benchmark/apptronik_apollo/scene_hfield.xml --nworld=4096 --nconmax=1500000 --njmax=512 --keyframe=0
```

this pr
```
Summary for 4096 parallel rollouts

Total JIT time: 0.57 s
Total simulation time: 26.83 s
Total steps per second: 152,693
Total realtime factor: 763.46 x
Total time per step: 6549.11 ns
```

main
```
Rolling out 1000 steps at dt = 0.005...
ncollision overflow - please increase nconmax to 1550266
```